+++ feature/receipt/build.gradle.kts
@@ -8,15 +8,23 @@
  * See https://github.com/openMF/mobile-wallet/blob/master/LICENSE.md
  */
 plugins {
-    alias(libs.plugins.mifospay.android.feature)
-    alias(libs.plugins.mifospay.android.library.compose)
+    alias(libs.plugins.mifospay.cmp.feature)
+    alias(libs.plugins.kotlin.parcelize)
 }
 
 android {
-    namespace = "org.mifospay.receipt"
+    namespace = "org.mifospay.feature.receipt"
 }
 
-dependencies {
-    // TODO:: this should be removed
-    implementation(libs.squareup.okhttp)
+kotlin {
+    sourceSets {
+        commonMain.dependencies {
+            implementation(compose.ui)
+            implementation(compose.foundation)
+            implementation(compose.material3)
+            implementation(compose.components.resources)
+            implementation(compose.components.uiToolingPreview)
+            implementation(libs.squareup.okio)
+        }
+    }
 }
