+++ feature/receipt/src/androidMain/AndroidManifest.xml
