+++ feature/receipt/src/commonMain/composeResources/drawable/feature_receipt_ic_download.xml
