+++ feature/receipt/src/commonMain/composeResources/drawable/feature_receipt_mifospay_round_logo.png
