+++ feature/receipt/src/commonMain/composeResources/values/strings.xml
