+++ mifospay-android/dependencies/prodReleaseRuntimeClasspath.tree.txt
@@ -2360,6 +2360,38 @@
 |    |    |         +--- org.jetbrains.kotlin:kotlin-stdlib:1.9.24 -> 2.0.20 (*)
 |    |    |         \--- org.jetbrains.kotlin:kotlin-reflect:1.9.24 -> 2.0.20 (*)
 |    |    \--- org.jetbrains.kotlin:kotlin-parcelize-runtime:2.0.20 (*)
+|    +--- project :feature:receipt
+|    |    +--- androidx.lifecycle:lifecycle-runtime-compose:2.8.6 (*)
+|    |    +--- androidx.lifecycle:lifecycle-viewmodel-compose:2.8.6 (*)
+|    |    +--- androidx.tracing:tracing-ktx:1.3.0-alpha02 (*)
+|    |    +--- io.insert-koin:koin-bom:4.0.0-RC2 (*)
+|    |    +--- io.insert-koin:koin-android:4.0.0-RC2 (*)
+|    |    +--- io.insert-koin:koin-androidx-compose:4.0.0-RC2 (*)
+|    |    +--- io.insert-koin:koin-androidx-navigation:4.0.0-RC2 (*)
+|    |    +--- io.insert-koin:koin-core-viewmodel:4.0.0-RC2 (*)
+|    |    +--- org.jetbrains.kotlin:kotlin-stdlib:2.0.20 (*)
+|    |    +--- io.insert-koin:koin-core:4.0.0-RC2 (*)
+|    |    +--- io.insert-koin:koin-annotations:1.4.0-RC4 (*)
+|    |    +--- project :core:ui (*)
+|    |    +--- project :core:designsystem (*)
+|    |    +--- project :core:data (*)
+|    |    +--- io.insert-koin:koin-compose:1.2.0-Beta4 -> 4.0.0-RC2 (*)
+|    |    +--- io.insert-koin:koin-compose-viewmodel:1.2.0-Beta4 -> 4.0.0-RC2 (*)
+|    |    +--- org.jetbrains.compose.runtime:runtime:1.7.0-rc01 (*)
+|    |    +--- org.jetbrains.androidx.lifecycle:lifecycle-viewmodel-compose:2.8.2 (*)
+|    |    +--- org.jetbrains.androidx.lifecycle:lifecycle-viewmodel:2.8.2 -> 2.8.3-rc01 (*)
+|    |    +--- org.jetbrains.androidx.lifecycle:lifecycle-viewmodel-savedstate:2.8.2 (*)
+|    |    +--- org.jetbrains.androidx.savedstate:savedstate:1.2.2 (*)
+|    |    +--- org.jetbrains.androidx.core:core-bundle:1.0.1 (*)
+|    |    +--- org.jetbrains.androidx.navigation:navigation-compose:2.8.0-alpha10 (*)
+|    |    +--- org.jetbrains.kotlinx:kotlinx-collections-immutable:0.3.8 (*)
+|    |    +--- org.jetbrains.compose.ui:ui:1.7.0-rc01 (*)
+|    |    +--- org.jetbrains.compose.foundation:foundation:1.7.0-rc01 (*)
+|    |    +--- org.jetbrains.compose.material3:material3:1.7.0-rc01 (*)
+|    |    +--- org.jetbrains.compose.components:components-resources:1.7.0-rc01 (*)
+|    |    +--- org.jetbrains.compose.components:components-ui-tooling-preview:1.7.0-rc01 (*)
+|    |    +--- com.squareup.okio:okio:3.9.1 (*)
+|    |    \--- org.jetbrains.kotlin:kotlin-parcelize-runtime:2.0.20 (*)
 |    +--- org.jetbrains.kotlin:kotlin-stdlib:2.0.20 (*)
 |    +--- io.insert-koin:koin-annotations:1.4.0-RC4 (*)
 |    +--- project :core:ui (*)
